rename `nextra/ssg` to `nextra/data`
